requirements.txt
aiohttp>=3.9.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
pyyaml>=6.0
openai>=1.30.0
zai-sdk>=0.2.2
structlog>=24.0.0
apscheduler>=3.10.0
trafilatura>=1.6.0
chardet>=5.0.0
aiosqlite>=0.19.0
duckduckgo-search>=6.0.0
tiktoken>=0.7.0
playwright==1.49.1
python-dateutil>=2.8.0
clickhouse-connect>=0.7.0
